Crate gfx_texture

A Gfx texture representation that works nicely with Piston libraries.

Modules

ops

Image operations for textures.

Structs

Texture

Represents a texture.

TextureContext

Context required to create and update textures.

TextureSettings

Texture creation parameters.

Enums

Error

Create creation or update error.

Filter

Sampling filter

Flip

Flip settings.

Format

Texture format.

Wrap

Wrap mode

Traits

CreateTexture

Implemented by textures for creation.

ImageSize

Implemented by all images to be used with generic algorithms.

TextureOp

Implemented by texture operations.

UpdateTexture

Implemented by textures for updating.